# api-gateway/src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: Order Processing System API
  description: |
    Unified API Gateway for the Order Processing System.

    **Authentication**: All endpoints (except health/swagger) require `Authorization: Bearer <JWT>` (RS256).
    **Idempotency**: `POST /api/orders` requires `Idempotency-Key` header (UUIDv4) to prevent duplicates.
    **Tracing**: Include `X-Trace-Id` header for distributed tracing. Generated automatically if omitted.
  version: "1.0.0"
  contact:
    name: OPS Engineering
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.ops.example.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NotificationResponse:
      type: object
      properties:
        id:        { type: string, format: uuid }
        orderId:   { type: string, format: uuid }
        channel:   { type: string, enum: [EMAIL, SMS, PUSH] }
        recipient: { type: string }
        template:  { type: string }
        status:    { type: string, enum: [PENDING, DELIVERED, FAILED] }
        attempts:  { type: integer }
        sentAt:    { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }

    # ── Error schema ───────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for an order
      operationId: listNotifications
      parameters:
        - name: orderId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/NotificationResponse" } } } } }
  /api/notifications/{id}:
    get:
      tags: [Notifications]
      summary: Get a notification by ID
      operationId: getNotification
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationResponse" } } } }
        "404": { description: Notification not found }